Level 1, Key, Locked Door added
I added Level 1, a Key, and a lockable door.  Level 1 is a good test of
these features.  It is simple, easy to pass, but allows you to get
introduced to the basic game mechanics.  The key has a script that
destroys itself and adds a term to the players "Item" list.  At present
its a string list so that will need to be upgraded.  Also, I'm now
finding a lot of issues with the movement code, particularly when
jumping.  I'll have to work with the movement system more.